feat: :sparkles: i created layout component and apply to app structure
i created a layout component in components/layout/layout.js, providing a consistent structure with a header, main container, and footer for the BotoCar app.
i styled the component using css modules and integrated it into the app by wrapping all pages within the layout in _app.js. this setup includes a header link to the homepage and basic information for the footer.