feat: :sparkles: i created dynamic filter route for car price filtering
i created a dynamic route file pages/filter/[...slug].js to handle price filtering for cars based on user input.
the FilteredCars component uses useRouter to retrieve the minimum and maximum price from the URL, then filters the carsData array accordingly.
if no cars match the filter criteria, a message is displayed to inform the user.
otherwise, the filtered list of cars is passed to the CarsList component for display.